Merge/public

+++ src/HLS/simulation/memory_initialization_writer.cpp
@@ -149,7 +149,7 @@ void MemoryInitializationWriter::Process(const std::string& content)
    }
    THROW_ASSERT(base_type, "");
    std::string binary_value = "";
-   unsigned int size = 0;
+   unsigned long long size = 0;
    switch(GET_CONST_NODE(base_type)->get_kind())
    {
       case integer_type_K:
