+++ src/HLS/virtual_components/adder_conn_obj.hpp
@@ -54,7 +54,7 @@
 class adder_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned int bitsize;
+   unsigned long long bitsize;
 
    /// when true the addition can trim the last bits
    bool is_aligned_adder_p;
@@ -79,7 +79,7 @@ class adder_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned int _bitsize)
+   void add_bitsize(unsigned long long _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -87,7 +87,7 @@ class adder_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned int get_bitsize() const
+   unsigned long long get_bitsize() const
    {
       return bitsize;
    }
